Fix DELETE and PUT requests being filetered out by Nginx server configuration
Introducing changes that address issue of DELETE and PUT requests on Nginx proxy server being rejected with 403 code due to misconfiguration or other security measures. Possibly solves the same issues with IIS server if requests are not mapped properly. Reference to issue posted on pkp ojs forum https://forum.pkp.sfu.ca/t/issue-with-nginx-and-apache-when-delete-and-put-requests/61019?u=ssalavat