pkp/pkp-lib#8880 Remove grid role and focus handling from Table component
After reading a lot about the difference between tables and grids,
I've decided that our tables are not grids and should not implement
the grid role specification.

There is some ambiguity about when a table becomes interactive
enough to qualify as a grid. According to the W3C examples, a simple
table with an interactive element can be a grid. However, a lot of
the advice from accessibility consultants is to use a table unless
building a spreadsheet-like app.

For that reason, I decided it is probably best to keep the Table
component as a simple table role. Although we have some interactive
elements in our tables, they are still primarily presenting tabular
data. Following the principle of no aria is better than bad aria,
it is probably least harmful to avoid the grid specification for now.