Corrects listing of Taiwan to reflect the actual name of the country
at present Taiwan is listed in dropdowns as "Taiwan, Province of China" as in the PRC/Mainland China. To scholars working in Taiwan and of Taiwanese heritage, this is insulting and offensive. This PR corrects the name of Taiwan to reflect the name used by people on the island of Taiwan.